.circle/requirements.txt
jsonschema==2.5.1
pyyaml==3.11
requests[security]==2.10.0
